+++ changelog.md
@@ -1,4 +1,12 @@
-1.0.0 (2016-07-02)
+1.0.1 (2016-08-16)
+------------------
+
+- Fix (Unknow call) when function is called
+- Add preview only when available
+- Show \Closure in File::Line column, instead of (Unknow file)::(Uknow line)
+
+1.0.0 (2016-08-02)
 ------------------
 
 - Create \DumpBacktrace class
+